Create check_reorg.sh
Improving availability during table and index maintenance

Reorganization operations (reorgs) on tables or indexes tend to be resource intensive activities which limit the availability of the tables and reduce concurrency. However, if  the source code listed above follows these general guidelines, you should be able to reduce the impact of reorg-related outages:

* Perform reorgs only if necessary
* Use only type-2 indexes
* Reduce the need to perform reorgs
* Minimize the impact of your reorg
* Performs index and table reorgs only if necessary 

Reorgs should not be performed periodically just as a matter of routine. In fact, many DB2 indexes and tables will seldom ever need to be reorganized.  The those that do this software will only to what is necessary.